move logic to data layer and rename
move logic of getteng bearer token by code usecase to data layer and rename it
now domain doesn't work with android context
