+++ myProject/app/src/main/java/com/blblblbl/myapplication/presentation/viewModel/MainViewModel.kt
@@ -2,15 +2,15 @@ package com.blblblbl.myapplication.presentation.viewModel
 
 import android.net.Uri
 import androidx.lifecycle.ViewModel
-import com.blblblbl.myapplication.domain.usecase.AuthUseCase
+import com.blblblbl.myapplication.domain.usecase.GetBearerTokenUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
 
 @HiltViewModel
 class MainViewModel @Inject constructor(
-    private val authUseCase: AuthUseCase
+    private val getBearerTokenUseCase: GetBearerTokenUseCase
 ):ViewModel(){
     fun saveAuthToken(uri:Uri){
-        authUseCase.execute(uri)
+        getBearerTokenUseCase.execute(uri)
     }
 }
